backend-python/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
psycopg2-binary==2.9.9
cryptography==41.0.7
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
